move OpentronsHTTPAPI class to new http_api module so its reusable use Point class from hardware project instead of recreating it add CalibrationTarget enum and use it to return the correct calibration offset polarity in find_calibration_structure_position add urllib3, certifi, and charset-normalizer to Pipfile.lock for hardware_testing module
