+++ api/src/opentrons/hardware_control/ot3_calibration.py
@@ -47,6 +47,11 @@ class CalibrationMethod(Enum):
     NONCONTACT_PASS = "noncontact pass"
 
 
+class CalibrationTarget(Enum):
+    DECK_OBJECT = "deck_object"
+    GANTRY_INSTRUMENT = "gantry_instrument"
+
+
 class CalibrationStructureNotFoundError(RuntimeError):
     def __init__(self, structure_height: float, lower_limit: float) -> None:
         super().__init__(
@@ -570,6 +575,7 @@ async def find_calibration_structure_position(
     mount: OT3Mount,
     nominal_center: Point,
     method: CalibrationMethod = CalibrationMethod.BINARY_SEARCH,
+    target: CalibrationTarget = CalibrationTarget.GANTRY_INSTRUMENT,
     raise_verify_error: bool = True,
 ) -> Point:
     """Find the calibration square offset given an arbitry postition on the deck."""
@@ -582,7 +588,15 @@ async def find_calibration_structure_position(
     found_center = await find_calibration_structure_center(
         hcapi, mount, initial_center, method, raise_verify_error
     )
-    return nominal_center - found_center
+
+    offset = nominal_center - found_center
+    # NOTE: If the calibration target is a deck object the polarity of the calibrated
+    #  offset needs to be reversed. This is because we are using the gantry instrument
+    #  to calibrate a stationary object on the deck and need to find the offset of that
+    #  deck object relative to the deck and not the instrument which sits above the deck.
+    if target == CalibrationTarget.DECK_OBJECT:
+         return offset * -1
+    return offset
 
 
 async def find_slot_center_binary_from_nominal_center(
@@ -782,11 +796,8 @@ async def calibrate_module(
             # from the nominal position so we dont have to alter any other part of the system.
             nominal_position = nominal_position - PREP_OFFSET_DEPTH
             offset = await find_calibration_structure_position(
-                hcapi, mount, nominal_position, method=CalibrationMethod.BINARY_SEARCH
+                hcapi, mount, nominal_position, method=CalibrationMethod.BINARY_SEARCH, target=CalibrationTarget.DECK_OBJECT
             )
-
-            # need to flip the polarity since modules are fixed on the deck
-            offset = offset * -1
             await hcapi.save_module_offset(module_id, mount, slot, offset)
             return offset
         finally:
